+++ R/statuser-package.R
@@ -16,6 +16,8 @@
 #' \itemize{
 #'   \item \code{\link{twolines}}: Two-lines test for U-shapes (Simonsohn 2018)
 #'   \item \code{\link{interprobe}}: Probe and visualize nonlinear interactions (Simonsohn 2024; Montealegre & Simonsohn 2026)
+#'   \item \code{\link{stimulus.plot}}: Stimulus plots for matched/treated-stimulus designs (Simonsohn, Montealegre, & Evangelidis 2024)
+#'   \item \code{\link{stimulus.beeswarm}}: Beeswarm plots for compared-stimulus designs (Simonsohn, Montealegre, & Evangelidis 2024)
 #' }
 #'
 #' @section Graphing:
@@ -24,8 +26,9 @@
 #'   \item \code{\link{plot_cdf}}: Plot empirical cumulative distribution functions (optional, by group)
 #'   \item \code{\link{plot_density}}: Plot density functions (optional, by group)
 #'   \item \code{\link{plot_freq}}: Plot frequency of observed values (optional, by group)
-#'   \item \code{\link{plot_means}}: Barplot of means with confidence intervals and (optionally) tests
+#'   \item \code{\link{plot_means}}: Barplot of means with confidence intervals and (optionally) tests. See also \code{\link{stimulus.plot}} for per-stimulus plots.
 #'   \item \code{\link{plot_gam}}: Plot fitted GAM values for a focal predictor
+#'   \item \code{\link{text2}}: like text() adding horizontal alignment and background color
 #' }
 #'
 #' @section Formatting:
@@ -41,7 +44,6 @@
 #'   \item \code{\link{clear}}: Clear environment, console, and all graphics devices
 #'   \item \code{\link{list2}}: Like list(), but unnamed objects are automatically named
 #'   \item \code{\link{convert_to_sql}}: Convert CSV files to SQL INSERT statements
-#'   \item \code{\link{text2}}: like text() adding text-alignment and background color
 
 #' }
 #'
@@ -56,6 +58,9 @@
 #' Simonsohn, U. (2024). Interacting with curves: How to validly test and probe interactions in the real (nonlinear) world. \emph{Advances in Methods and Practices in Psychological Science}, 7(1), Article 25152459231207787. \doi{10.1177/25152459231207787}
 #'
 #' Montealegre, D., & Simonsohn, U. (2026). \emph{Johnson Neyman 2.0} (working paper).
+#'
+#' Simonsohn, U., Montealegre, A., & Evangelidis, I. (2024). Stimulus sampling reimagined.
+#' \emph{Advances in Methods and Practices in Psychological Science}.
 #' 
 #' @seealso
 #' Useful links: